// apps/server/src/config/app-env.d.ts
export declare const appEnv: {
    PORT: string;
    DATABASE_URL: string;
    MOBILE_CLIENT: string;
    WEB_CLIENT: string;
    BASE_PATH: string;
    JWT: {
        SECRET: string;
        EXPIRES_IN: string;
        REFRESH_SECRET: string;
        REFRESH_EXPIRES_IN: string;
        RESET_PASSWORD_SECRET: string;
        RESET_PASSWORD_EXPIRES_IN: string;
    };
};
export default appEnv;